+++ src/Polly/Timeout/AsyncTimeoutEngine.cs
@@ -47,7 +47,9 @@ namespace Polly.Timeout
                     }
                     catch (Exception ex)
                     {
-                        if (timeoutCancellationTokenSource.IsCancellationRequested)
+                        // Note that we cannot rely on testing (operationCanceledException.CancellationToken == combinedToken || operationCanceledException.CancellationToken == timeoutCancellationTokenSource.Token)
+                        // as either of those tokens could have been onward combined with another token by executed code, and so may not be the token expressed on operationCanceledException.CancellationToken.
+                        if (ex is OperationCanceledException && timeoutCancellationTokenSource.IsCancellationRequested)
                         {
                             await onTimeoutAsync(context, timeout, actionTask, ex).ConfigureAwait(continueOnCapturedContext);
                             throw new TimeoutRejectedException("The delegate executed asynchronously through TimeoutPolicy did not complete within the timeout.", ex);